added wolfram parser

+++ app/src/main/java/information/NutritionInfo.java
@@ -2,9 +2,9 @@ package information;
 
 public class NutritionInfo {
     private String name;
-    private long calories;
-    private long fiber;
-    private long protein;
+    private double calories;
+    private double fiber;
+    private double protein;
 
     public String getName() {
         return name;
@@ -14,27 +14,27 @@ public class NutritionInfo {
         this.name = name;
     }
 
-    public long getCalories() {
+    public double getCalories() {
         return calories;
     }
 
-    public void setCalories(long calories) {
+    public void setCalories(double calories) {
         this.calories = calories;
     }
 
-    public long getFiber() {
+    public double getFiber() {
         return fiber;
     }
 
-    public void setFiber(long fiber) {
+    public void setFiber(double fiber) {
         this.fiber = fiber;
     }
 
-    public long getProtein() {
+    public double getProtein() {
         return protein;
     }
 
-    public void setProtein(long protein) {
+    public void setProtein(double protein) {
         this.protein = protein;
     }
 }
